add util function getSigner
