+++ Project/scripts/castVote.ts
@@ -1,13 +1,9 @@
+/* eslint-disable node/no-missing-import */
 import { Contract, ethers } from "ethers";
 import "dotenv/config";
 import * as ballotJson from "../artifacts/contracts/CustomBallot.sol/CustomBallot.json";
-// eslint-disable-next-line node/no-missing-import
+import { getSigner } from "./util/getSigner";
 import { CustomBallot } from "../typechain";
-
-// This key is already public on Herong's Tutorial Examples - v1.03, by Dr. Herong Yang
-// Do never expose your keys like this
-const EXPOSED_KEY =
-  "8da4ef21b864d2cc526dbdb2a120bd2874c36c9d0a1fb7f8c63d7f7a8b41de8f";
 
 // TODO: Test once ballot contract is deployed
 
@@ -24,19 +20,7 @@ Script arguments:
 The voter is the address of the wallet in the environment.
 */
 async function main() {
-    const wallet =
-      process.env.MNEMONIC && process.env.MNEMONIC.length > 0
-        ? ethers.Wallet.fromMnemonic(process.env.MNEMONIC)
-        : new ethers.Wallet(process.env.PRIVATE_KEY ?? EXPOSED_KEY);
-    console.log(`Using address ${wallet.address}`);
-    const provider = ethers.providers.getDefaultProvider("goerli");
-    const signer = wallet.connect(provider);
-    const balanceBN = await signer.getBalance();
-    const balance = Number(ethers.utils.formatEther(balanceBN));
-    console.log(`Wallet balance ${balance}`);
-    if (balance < 0.01) {
-      throw new Error("Not enough ether");
-    }
+    const signer = await getSigner();
 
     if (process.argv.length < 3) throw new Error("Ballot address missing");
     const ballotAddress = process.argv[2];
