+++ Project/scripts/deployToken.ts
@@ -1,43 +1,12 @@
 import { ethers } from "ethers";
 import "dotenv/config";
-import * as tokenJson from "../artifacts/contracts/Token.sol/MyToken.json"
-
-// This key is already public on Herong's Tutorial Examples - v1.03, by Dr. Herong Yang
-// Do never expose your keys like this
-const EXPOSED_KEY =
-  "8da4ef21b864d2cc526dbdb2a120bd2874c36c9d0a1fb7f8c63d7f7a8b41de8f";
-
-function setupProvider() {
-  const infuraOptions = process.env.INFURA_API_KEY
-    ? process.env.INFURA_API_SECRET
-      ? {
-        projectId: process.env.INFURA_API_KEY,
-        projectSecret: process.env.INFURA_API_SECRET,
-      }
-      : process.env.INFURA_API_KEY
-    : "";
-  const options = {
-    alchemy: process.env.ALCHEMY_API_KEY,
-    infura: infuraOptions,
-  };
-  const provider = ethers.providers.getDefaultProvider("goerli", options);
-  return provider;
-}
+import * as tokenJson from "../artifacts/contracts/Token.sol/MyToken.json";
+// eslint-disable-next-line node/no-missing-import
+import { getSigner } from "./util/getSigner";
 
 async function main() {
-  const wallet =
-    process.env.MNEMONIC && process.env.MNEMONIC.length > 0
-      ? ethers.Wallet.fromMnemonic(process.env.MNEMONIC)
-      : new ethers.Wallet(process.env.PRIVATE_KEY ?? EXPOSED_KEY);
-  console.log(`Using address ${wallet.address}`);
-  const provider = setupProvider();
-  const signer = wallet.connect(provider);
-  const balanceBN = await signer.getBalance();
-  const balance = Number(ethers.utils.formatEther(balanceBN));
-  console.log(`Wallet balance ${balance}`);
-  if (balance < 0.01) {
-    throw new Error("Not enough ether");
-  }
+  const signer = await getSigner();
+
   console.log("Deploying Token contract");
   const tokenFactory = new ethers.ContractFactory(
     tokenJson.abi,
@@ -49,7 +18,7 @@ async function main() {
   await tokenContract.deployed();
   console.log("Completed");
   console.log(`Contract deployed at ${tokenContract.address}`);
-  const mintTx = await tokenContract.mint(wallet.address, 100);
+  const mintTx = await tokenContract.mint(signer.address, 100);
   await mintTx.wait();
 }
